real_time_quotes_consumer/requirements.txt
Django>=3.2,<4.0
djangorestframework
psycopg2-binary
kafka-python
line_profiler
httpx
python-dotenv
channels  
channels_redis
daphne
redis